Renamed utils directory to src/lib
