+++ src/lib/slpMiddleware.js
